chore: moved more hooks
