+++ apps/easypid/src/features/share/FunkeRequestedAttributesDetailScreen.tsx
@@ -1,4 +1,3 @@
-import { type CredentialForDisplayId, useCredentialForDisplayById } from '@package/agent'
 import {
   CredentialAttributes,
   FunkeCredentialCard,
@@ -23,13 +22,14 @@ import {
   useToastController,
 } from '@package/ui'
 import { metadataForDisplay } from '@paradym/wallet-sdk/src/display/common'
+import { type CredentialId, useCredentialById } from '@paradym/wallet-sdk/src/hooks/useCredentialById'
 import { useRouter } from 'expo-router'
 import { useRef, useState } from 'react'
 import { FadeInUp, FadeOutUp } from 'react-native-reanimated'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 interface FunkeRequestedAttributesDetailScreenProps {
-  id: CredentialForDisplayId
+  id: CredentialId
   disclosedPayload: Record<string, unknown>
   disclosedAttributeLength: number
 }
@@ -42,7 +42,7 @@ export function FunkeRequestedAttributesDetailScreen({
   const toast = useToastController()
   const { handleScroll, isScrolledByOffset, scrollEventThrottle } = useScrollViewPosition()
   const { bottom } = useSafeAreaInsets()
-  const { credential: activeCredential, isLoading } = useCredentialForDisplayById(id)
+  const { credential: activeCredential, isLoading } = useCredentialById(id)
   const router = useRouter()
   const [scrollViewHeight, setScrollViewHeight] = useState(0)
   const { withHaptics } = useHaptics()
