+++ apps/easypid/src/features/wallet/components/InboxIcon.tsx
@@ -1,7 +1,7 @@
-import { useHasInboxNotifications } from '@package/agent'
 import { useHaptics } from '@package/app'
 import { AnimatedStack, Circle, IconContainer, Stack } from '@package/ui'
 import { HeroIcons } from '@package/ui'
+import { useHasInboxNotifications } from '@paradym/wallet-sdk/src/hooks/useHasInboxNotifications'
 import { useRouter } from 'expo-router'
 import { ZoomIn, ZoomOut } from 'react-native-reanimated'
