+++ apps/easypid/src/features/wallet/components/LatestActivityCard.tsx
@@ -1,8 +1,8 @@
 import { useActivities } from '@easypid/features/activity/activityRecord'
-import { useCredentialsForDisplay } from '@package/agent'
 import { useHaptics } from '@package/app'
 import { InfoButton } from '@package/ui'
 import { formatRelativeDate } from '@package/utils'
+import { useCredentials } from '@paradym/wallet-sdk/src/hooks/useCredentials'
 import { useRouter } from 'expo-router'
 import { useMemo } from 'react'
 
@@ -11,7 +11,7 @@ export function LatestActivityCard() {
   const { withHaptics } = useHaptics()
   const { activities } = useActivities()
   const latestActivity = activities[0]
-  const { credentials } = useCredentialsForDisplay()
+  const { credentials } = useCredentials()
 
   const pushToActivity = withHaptics(() => push('/activity'))
