+++ packages/agent/src/index.ts
@@ -18,7 +18,6 @@ export {
 export * from './batch'
 export * from './invitation'
 export * from './hooks'
-export * from './mediation'
 export * from './crypto'
 export * from './storage'
 export { migrateLegacyParadymWallet } from './migrateLegacyParadymWallet'
