+++ packages/app/src/features/notifications/NotificationInboxScreen.tsx
@@ -1,7 +1,6 @@
-import { useInboxNotifications } from '@package/agent'
 import { Heading, ScrollView, Stack, YStack } from '@package/ui'
+import { useInboxNotifications } from '@paradym/wallet-sdk/src/hooks/useInboxNotifications'
 import { useRouter } from 'expo-router'
-
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { InboxNotificationRowCard, NoContentInbox } from '../../components'
 import { useScrollViewPosition } from '../../hooks'
