--- packages/sdk/src/hooks/useCredentialsById.tsx
@@ -1,10 +1,0 @@
-import { useCredentials } from './useCredentials'
-
-export const useCredentialsById = (id: string) => {
-  const { isLoading, credentials } = useCredentials()
-
-  return {
-    isLoading,
-    credential: credentials.find((credential) => credential.id === id),
-  }
-}
